+++ internal/server/handlers.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 	"strings"
 
-	"git.hq.quigley.com/products/ranger/internal/api"
-	"git.hq.quigley.com/products/ranger/internal/model"
-	"git.hq.quigley.com/products/ranger/internal/workspace"
+	"github.com/michaelquigley/ranger/internal/api"
+	"github.com/michaelquigley/ranger/internal/model"
+	"github.com/michaelquigley/ranger/internal/workspace"
 )
 
 func (s *Server) GetBoard(_ context.Context) (*api.Board, error) {